+++ crates/backend/Cargo.toml
@@ -10,10 +10,12 @@ edition = "2018"
 crate-type = ["cdylib", "rlib"]
 
 [dependencies]
-strsim = "0.10.0"
-serde_ignored = "0.1.2"
-serde = { version = "1.0.114", features = ["derive"] }
 cargo_metadata = "0.10.0"
-wasm-bindgen-cli-support = "0.2.63"
-toml = "0.5.6"
 notify = "4.0.15"
+serde = { version = "1.0.114", features = ["derive"] }
+serde_ignored = "0.1.2"
+strsim = "0.10.0"
+toml = "0.5.6"
+tokio = { version = "0.2", features = ["macros"] }
+warp = "0.2.3"
+wasm-bindgen-cli-support = "0.2.63"
