+++ examples/hello-world/Cargo.toml
@@ -10,6 +10,6 @@ edition = "2018"
 crate-type = ["cdylib"]
 
 [dependencies]
-wasm-bindgen = "0.2.63"
-elvis-web = { path = "../../crates/web" }
 elvis = { path = "../.." }
+elvis-web = { path = "../../crates/web" }
+wasm-bindgen = "0.2.63"
